hid_class: add support for custom subclass and class
Useful for emulating devices like Xbox 360 controllers, which are
essentialy HID devices, but present a nonstandard class and
subclass because Microsoft. I use this in
https://github.com/Mstrodl/devcade-controller with the linux xpad
driver and it behaves well, which is all I really needed it to do.

If this nonstandard functionality is not desired in this repo,
that's okay too :)